Split PageManager/elements/PageEditor.py into PageEditorParts/
